Standard function interface for quality candidate filters.
